doing: 整理组件关系
