+++ src/components/components.ts
@@ -1,4 +1,22 @@
-export { ReadHeader } from './ReadHeader';
-export { ArticleView } from './ArticleView';
-export { default as LoadingState } from './LoadingState';
-export { default as EmptyState } from './EmptyState';
+// 通用组件
+export { LoadingState } from './read/LoadingState';
+export { TableOfContents } from './read/TableOfContents';
+export { SourceForm } from './readme/SourceForm';
+
+// 介绍和设置相关组件
+export { Introduction } from './readme/Introduction';
+
+// 阅读视图相关组件
+export { ReadSidebar } from './read/ReadSidebar';
+export { ReadHeader } from './read/ReadHeader';
+export { EmptyState } from './read/EmptyState';
+export { Read } from './read/Read';
+export { HeadingBlockView } from './read/HeadingBlockView';
+
+// 画廊相关组件
+export { ArticleView } from './gallery/ArticleView';
+export { ArticleCard } from './gallery/ArticleCard';
+export { Gallery } from './gallery/Gallery';
+
+// 侧边栏
+export { Sidebar } from './Sidebar';
