+++ src/components/gallery/ArticleView.tsx
@@ -1,8 +1,8 @@
-import React, { useRef, useEffect } from 'react';
-import { RSSItem, ContentBlock, Article } from '../types';
-import TableOfContents from './TableOfContents';
-import { ContentBlockView } from './ContentBlock';
 import { setIcon } from 'obsidian';
+import React, { useEffect, useRef } from 'react';
+import { ContentBlockView } from '../read/ContentBlock';
+import { Article, ContentBlock, RSSItem } from '../../types';
+import TableOfContents from '../read/TableOfContents';
 
 // 统一接口，支持RSSItem和Article两种类型
 interface ArticleViewProps {
