+++ src/components/read/HeadingBlockView.tsx
@@ -1,13 +1,13 @@
 import React, { useState, useRef, useEffect } from 'react';
 import { setIcon } from 'obsidian';
-import { ContentBlock } from '../types';
+import { ContentBlock } from '../../types';
 
-interface BlockquoteProps {
+interface HeadingBlockProps {
     block: ContentBlock;
     onFavorite: (content: string, blockId: number) => void;
 }
 
-export const BlockquoteView: React.FC<BlockquoteProps> = ({ block, onFavorite }) => {
+export const HeadingBlockView: React.FC<HeadingBlockProps> = ({ block, onFavorite }) => {
     const [isHovered, setIsHovered] = useState(false);
     const [isFavorited, setIsFavorited] = useState(false);
     const favoriteButtonRef = useRef<HTMLButtonElement>(null);
@@ -23,14 +23,17 @@ export const BlockquoteView: React.FC<BlockquoteProps> = ({ block, onFavorite })
         onFavorite(block.content, block.id);
     };
     
+    // 动态渲染不同级别的标题
+    const HeadingTag = `h${block.level || 2}` as keyof JSX.IntrinsicElements;
+    
     return (
         <div 
-            className={`blockquote-block block ${isHovered ? 'hovered' : ''} ${isFavorited ? 'favorited' : ''}`}
+            className={`heading-block block level-${block.level} ${isHovered ? 'hovered' : ''} ${isFavorited ? 'favorited' : ''}`}
             onMouseEnter={() => setIsHovered(true)}
             onMouseLeave={() => setIsHovered(false)}
         >
-            <blockquote 
-                className="blockquote-content"
+            <HeadingTag 
+                className={`heading-content heading-${block.level}`}
                 dangerouslySetInnerHTML={{ __html: block.content }}
             />
