+++ src/views/readme-view.ts
@@ -2,7 +2,7 @@ import { ItemView, WorkspaceLeaf } from "obsidian";
 import { VIEW_TYPES } from "../types";
 import { createRoot } from 'react-dom/client';
 import React from 'react';
-import { ReadMe } from '../components/ReadMe';
+import { ReadMe } from '../components/readme/ReadMe';
 import type RSSFlowPlugin from '../main';
 import { i18n } from '../i18n';
